Move Firebase Admin Utils to Shared (#171)
* Move Firebase Admin Utils to Shared

* Prettified Code!
