+++ .github/workflows/shared.yml
@@ -1,0 +1,39 @@
+name: Shared
+on:
+  pull_request:
+  push:
+    branches:
+      - main
+
+jobs:
+  check_applicability:
+    name: Check Applicability
+    runs-on: ubuntu-latest
+    outputs:
+      should_skip: ${{ steps.skip_check.outputs.should_skip }}
+    steps:
+      - id: skip_check
+        uses: fkirc/skip-duplicate-actions@v5
+        with:
+          paths: '["shared/**", ".github/workflows/shared.yml"]'
+          skip_after_successful_duplicate: "false"
+
+  test_deploy:
+    name: Test Shared Module
+    needs: check_applicability
+    if: needs.check_applicability.outputs.should_skip != 'true'
+    runs-on: ubuntu-latest
+    steps:
+      - name: Checkout Repo
+        uses: actions/checkout@v2
+
+      - name: Install & Setup Firebase Emulator Suite
+        run: |
+          npm -g install firebase-tools && \
+          firebase setup:emulators:firestore
+
+      - name: Install Dependencies
+        run: npm ci
+
+      - name: Run Tests
+        run: npm run shared:test
