+++ admin/src/App.tsx
@@ -9,7 +9,7 @@ import {
 	performAlgoliaTextSearch,
 	User,
 } from '@camberi/firecms';
-import { AdminUser } from '@socialincome/shared/types';
+import { AdminUser } from '../../shared/src/types';
 import { connectAuthEmulator, getAuth } from 'firebase/auth';
 import { connectFirestoreEmulator, getFirestore } from 'firebase/firestore';
 import { connectStorageEmulator, getStorage } from 'firebase/storage';
