+++ admin/src/collections/Contributions.ts
@@ -1,10 +1,5 @@
 import { buildCollection, buildProperties } from '@camberi/firecms';
-import {
-	Contribution,
-	ContributionSourceKey,
-	CONTRIBUTION_FIRESTORE_PATH,
-	StatusKey,
-} from '@socialincome/shared/types';
+import { Contribution, ContributionSourceKey, CONTRIBUTION_FIRESTORE_PATH, StatusKey } from '../../../shared/src/types';
 
 export const contributionsCollection = buildCollection<Contribution>({
 	name: 'Contributions',
