+++ admin/src/collections/Recipients.tsx
@@ -6,8 +6,8 @@ import {
 	buildProperty,
 	StringPropertyPreview,
 } from '@camberi/firecms';
-import { PARTNER_ORGANISATION_FIRESTORE_PATH, Recipient, RECIPIENT_FIRESTORE_PATH } from '@socialincome/shared/types';
-import { getMonthIDs } from '@socialincome/shared/utils';
+import { PARTNER_ORGANISATION_FIRESTORE_PATH, Recipient, RECIPIENT_FIRESTORE_PATH } from '../../../shared/src/types';
+import { getMonthIDs } from '../../../shared/utils';
 import { isUndefined } from 'lodash';
 import { BuildCollectionProps } from './index';
 import { paymentsCollection, paymentStatusMap } from './Payments';
