+++ functions/tests/etl/importPostfinanceBalance.test.ts
@@ -1,8 +1,8 @@
 import { describe, expect, test } from '@jest/globals';
 import firebaseFunctionsTest from 'firebase-functions-test';
-import { BankBalance, BANK_BALANCE_FIRESTORE_PATH, getIdFromBankBalance } from '../../../shared/types';
-import { doc } from '../useFirestoreAdmin';
-import * as importPostfinanceBalance from './importPostfinanceBalance';
+import { BankBalance, BANK_BALANCE_FIRESTORE_PATH, getIdFromBankBalance } from '../../../shared/src/types';
+import { doc } from '../../../shared/src/firebase/firestoreAdmin';
+import * as importPostfinanceBalance from '../../src/etl/importPostfinanceBalance';
 const { cleanup } = firebaseFunctionsTest();
 
 describe('importPostfinanceBalance', () => {
