+++ shared/jest.config.js
@@ -1,0 +1,7 @@
+/** @type {import('ts-jest/dist/types').InitialOptionsTsJest} */
+module.exports = {
+	preset: 'ts-jest',
+	testEnvironment: 'node',
+	testPathIgnorePatterns: ['.d.ts', '.js'],
+	testTimeout: 60000,
+};
