+++ shared/package.json
@@ -2,7 +2,20 @@
 	"name": "@socialincome/shared",
 	"version": "0.1.0",
 	"private": true,
+	"scripts": {
+		"test": "firebase emulators:exec --only firestore,storage --project social-income-prod --config ../firebase.json --import ../seed 'npm run test:app'",
+		"test:app": "jest --forceExit --roots tests/ src/"
+	},
+	"dependencies": {
+		"@jest/globals": "^29.1.2",
+		"@types/jest": "^29.1.1",
+		"firebase-functions-test": "^2.4.0",
+		"jest": "^29.1.2",
+		"ts-jest": "^29.0.3",
+		"firebase-admin": "^11.1.0"
+	},
 	"devDependencies": {
+		"firebase-functions-test": "^2.4.0",
 		"@camberi/firecms": "2.0.0-alpha.37"
 	}
 }
