+++ shared/src/types/admin/NewsletterSubscriber.ts
