+++ shared/tsconfig.json
@@ -1,0 +1,19 @@
+{
+	"compilerOptions": {
+		"allowSyntheticDefaultImports": true,
+		"esModuleInterop": true,
+		"lib": ["dom", "es2020"],
+		"module": "commonjs",
+		"moduleResolution": "node",
+		"noImplicitReturns": true,
+		"noUnusedLocals": true,
+		"outDir": "dist",
+		"preserveConstEnums": true,
+		"resolveJsonModule": true,
+		"strict": true,
+		"skipLibCheck": true,
+		"target": "es2017",
+		"types": ["jest", "node"]
+	},
+	"include": ["/**/*"]
+}
